backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.7.0
httpx==0.27.0
redis==5.0.8
python-multipart==0.0.9
pillow==10.4.0
scikit-learn==1.5.1
statsmodels==0.14.2
numpy==1.26.4
pandas==2.2.2
pydantic==2.9.0
